+++ changelog.md
@@ -1,5 +1,11 @@
 # Gro changelog
 
+## 0.9.2
+
+- fix `src/build.task.ts` ([#133](https://github.com/feltcoop/gro/pull/133)):
+  - forward dev arg to esbuild options builder
+  - fix return values of rollup plugins to respect sourcemaps
+
 ## 0.9.1
 
 - fix `src/build.task.ts` to correctly use `sourcemap` and `target` Gro config options